Adding a demo notebook to demonstrate PoC on running dask cluster.
